♻️ Simplify ESLint config and fix newly enabled rule violations
Remove 16 redundant rule overrides from eslint.config.mjs:
- 6 deprecated/redundant/no-op rules (interface-name-prefix, no-empty-interface,
  explicit-function-return-type, no-unsafe-function-type, no-wrapper-object-types,
  no-restricted-types, no-var-requires)
- 3 rules re-enabled from preset with code fixes (no-unnecessary-type-assertion,
  restrict-template-expressions, restrict-plus-operands)

Fix violations for newly enabled rules:
- Remove 11 unnecessary type assertions (auto-fixed)
- Wrap template expressions with String() for never/generic types
- Compute string concatenations before type guard narrowing in WordsToLorem

+++ packages/fast-check/src/arbitrary/bigInt.ts
@@ -53,7 +53,7 @@ function extractBigIntConstraints(args: [] | [bigint, bigint] | [BigIntConstrain
     return constraints;
   } // args.length > 1
 
-  return { min: args[0] as bigint, max: args[1] as bigint }; // bigInt(min, max)
+  return { min: args[0] as bigint, max: args[1] }; // bigInt(min, max)
 }
 
 /**
